+++ lib/cassandra.rb
@@ -10,7 +10,7 @@ here = File.expand_path(File.dirname(__FILE__))
 
 class CassandraOld ; end
 unless CassandraOld.respond_to?(:VERSION)
-  require "#{here}/cassandra/0.8"
+  require "#{here}/cassandra_old/0.8"
 end
 
 $LOAD_PATH << "#{here}/../vendor/#{CassandraOld.VERSION}/gen-rb"
@@ -18,27 +18,27 @@ require "#{here}/../vendor/#{CassandraOld.VERSION}/gen-rb/cassandra"
 
 $LOAD_PATH << "#{here}"
 
-require 'cassandra/helpers'
-require 'cassandra/array'
-require 'cassandra/time'
-require 'cassandra/comparable'
-require 'cassandra/long'
-require 'cassandra/composite'
-require 'cassandra/dynamic_composite'
-require 'cassandra/ordered_hash'
-require 'cassandra/columns'
-require "#{here}/cassandra/protocol"
-require 'cassandra/batch'
-require "cassandra/#{CassandraOld.VERSION}/columns"
-require "cassandra/#{CassandraOld.VERSION}/protocol"
-require "cassandra/cassandra"
-require "cassandra/#{CassandraOld.VERSION}/cassandra"
+require 'cassandra_old/helpers'
+require 'cassandra_old/array'
+require 'cassandra_old/time'
+require 'cassandra_old/comparable'
+require 'cassandra_old/long'
+require 'cassandra_old/composite'
+require 'cassandra_old/dynamic_composite'
+require 'cassandra_old/ordered_hash'
+require 'cassandra_old/columns'
+require 'cassandra_old/protocol'
+require 'cassandra_old/batch'
+require "cassandra_old/#{CassandraOld.VERSION}/columns"
+require "cassandra_old/#{CassandraOld.VERSION}/protocol"
+require "cassandra_old/cassandra"
+require "cassandra_old/#{CassandraOld.VERSION}/cassandra"
 unless CassandraOld.VERSION.eql?("0.6")
-  require "cassandra/column_family"
-  require "cassandra/keyspace"
+  require "cassandra_old/column_family"
+  require "cassandra_old/keyspace"
 end
-require 'cassandra/constants'
-require 'cassandra/debug' if ENV['DEBUG']
+require 'cassandra_old/constants'
+require 'cassandra_old/debug' if ENV['DEBUG']
 
 begin
   require "cassandra_native"
