+++ lib/cassandra_old/0.6/columns.rb
