+++ lib/cassandra_old/0.7/columns.rb
