+++ lib/cassandra_old/0.7/protocol.rb
