+++ lib/cassandra_old/0.8/columns.rb
