+++ lib/cassandra_old/1.0.rb
