+++ lib/cassandra_old/1.1.rb
