+++ lib/cassandra_old/1.1/cassandra.rb
