+++ lib/cassandra_old/1.2.rb
