+++ lib/cassandra_old/1.2/columns.rb
