+++ lib/cassandra_old/array.rb
