+++ lib/cassandra_old/batch.rb
