+++ lib/cassandra_old/comparable.rb
