+++ lib/cassandra_old/composite.rb
