+++ lib/cassandra_old/constants.rb
