+++ lib/cassandra_old/dynamic_composite.rb
