+++ lib/cassandra_old/helpers.rb
