+++ lib/cassandra_old/keyspace.rb
