+++ lib/cassandra_old/long.rb
