+++ lib/cassandra_old/mock.rb
