+++ lib/cassandra_old/ordered_hash.rb
