+++ lib/cassandra_old/time.rb
