+++ test/test_helper.rb
@@ -1,7 +1,7 @@
 CASSANDRA_VERSION = ENV['CASSANDRA_VERSION'] || '0.8' unless defined?(CASSANDRA_VERSION)
 
 require 'test/unit'
-require "#{File.expand_path(File.dirname(__FILE__))}/../lib/cassandra/#{CASSANDRA_VERSION}"
+require "#{File.expand_path(File.dirname(__FILE__))}/../lib/cassandra_old/#{CASSANDRA_VERSION}"
 begin; require 'ruby-debug'; rescue LoadError; end
 
 begin
